Created a simple worm gate, segment and sender
The wormgate (wurmtor) listens for HTTP POSTs that contain a go file it
should execute. When it receives a POST request it writes the body (which it
assumes is a go file with a main func) to a file and executes the "go run"
command to execute whatever was in the file.

The "segment" appends text strings to a file named *output*.

The sender is responsible for grabbing the hello-world.go file and sending it
to the wurmtur using a HTTP POST request.